Added RunAction and SteppingAction
RunAction: defines one "run" in the simulation. defines dose, calculates it, and displays it add the end of the run

SteppingAction: defines the process in one "step" of the program, i.e. firing a particle, figuring out if it reaches the volume for which dose is being measured, and if yes adding the energy deposit to the total