Event action (#6)
* Add files via upload

* Create BTActionInitialzation.cc

* Create BTActionInitialzation.hh

* Update BTDetectorConstruction.cpp

* Update BTDetectorConstruction.hh

* Update BTActionInitialzation.hh

* Update BTPrimaryGeneratorAction.cpp

Defines the types of particles that can be used in the application. In this case, there's only 1: an 18 MeV gamma.

* Update BTPrimaryGeneratorAction.hh

* Added RunAction and SteppingAction

RunAction: defines one "run" in the simulation. defines dose, calculates it, and displays it add the end of the run

SteppingAction: defines the process in one "step" of the program, i.e. firing a particle, figuring out if it reaches the volume for which dose is being measured, and if yes adding the energy deposit to the total

* Event Action

* removed unrelated files

* fixed spook and header guard &
comment changes

* comment style change

* delete lingering comment

Co-authored-by: Delaney Dyment <70232125+dldyment@users.noreply.github.com>
Co-authored-by: Dhruv581 <66338432+Dhruv581@users.noreply.github.com>
Co-authored-by: pcmodeactivate <349348193@gapps.yrdsb.ca>
Co-authored-by: Kyle Tam <43300149+kylemtam@users.noreply.github.com>